Strip out most of the legacy save loading code.
