+++ src/main_menu.cpp
@@ -426,7 +426,6 @@ bool game::opening_screen()
                         werase(w_background);
                         wrefresh(w_background);
 
-                        MAPBUFFER.load(world->world_name);
                         start_game(world->world_name);
                         start = true;
                     } else if (sel2 == 1) {
@@ -668,7 +667,6 @@ bool game::opening_screen()
                         WORLDPTR world = world_generator->all_worlds[world_generator->all_worldnames[sel2]];
                         world_generator->set_active_world(world);
                         setup();
-                        MAPBUFFER.load(world->world_name);
 
                         load(world->world_name, savegames[sel3]);
                         start = true;
@@ -803,7 +801,6 @@ bool game::opening_screen()
                     }
                     werase(w_background);
                     wrefresh(w_background);
-                    MAPBUFFER.load(world_generator->active_world->world_name);
                     start_game(world_generator->active_world->world_name);
                     start = true;
                 }
